+++ pycm/pycm_handler.py
@@ -93,10 +93,16 @@ def __overall_stat_init__(cm):
     cm.SOA4 = cm.overall_stat["SOA4(Cicchetti)"]
     cm.Kappa = cm.overall_stat["Kappa"]
     cm.Overall_ACC = cm.overall_stat["Overall ACC"]
+    cm.TNR_Macro = cm.overall_stat["TNR Macro"]
     cm.TPR_Macro = cm.overall_stat["TPR Macro"]
+    cm.FNR_Macro = cm.overall_stat["FNR Macro"]
+    cm.FPR_Macro = cm.overall_stat["FPR Macro"]
     cm.PPV_Macro = cm.overall_stat["PPV Macro"]
     cm.ACC_Macro = cm.overall_stat["ACC Macro"]
+    cm.TNR_Micro = cm.overall_stat["TNR Micro"]
+    cm.FPR_Micro = cm.overall_stat["FPR Micro"]
     cm.TPR_Micro = cm.overall_stat["TPR Micro"]
+    cm.FNR_Micro = cm.overall_stat["FNR Micro"]
     cm.PPV_Micro = cm.overall_stat["PPV Micro"]
     cm.F1_Macro = cm.overall_stat["F1 Macro"]
     cm.F1_Micro = cm.overall_stat["F1 Micro"]
@@ -188,7 +194,10 @@ def __obj_assign_handler__(cm, matrix_param):
         CEN_dict=statistic_result["CEN"],
         MCEN_dict=statistic_result["MCEN"],
         AUC_dict=statistic_result["AUC"],
-        ICSI_dict=statistic_result["ICSI"])
+        ICSI_dict=statistic_result["ICSI"],
+        TNR=statistic_result["TNR"],
+        TN=statistic_result["TN"],
+        FP=statistic_result["FP"])
 
 
 def __obj_file_handler__(cm, file):